requirements.txt
torch
transformers
streamlit
pillow
pandas
numpy
diffusers
accelerate
safetensors
torchvision
